Model/requirements.txt
kafka-python
sentence-transformers
faiss-cpu
